backend/requirements.txt

fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.12.1
pydantic==2.5.0
pydantic-settings==2.0.3
python-jose==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
torch==2.1.0
torchvision==0.16.0
transformers==4.35.0
Pillow==10.1.0
numpy==1.26.2
pandas==2.1.3
reportlab==4.0.7
aiofiles==23.2.1
python-dotenv==1.0.0
slowapi==0.1.9
jinja2==3.1.2
